chore: minor changes

+++ src/routes/mod.rs
@@ -6,6 +6,6 @@ pub mod ping;
 #[derive(Serialize)]
 pub struct ResponseRequest<T> {
     message: Option<String>,
-    status: i32,
+    status: u16,
     data: T
 }
